+++ docs/changelog.md
@@ -20,7 +20,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 
 ### Added
-
+- container component
 - icon support for [panel component](components/BulmaPanel.md)
 
 ### Changed